tasks/now-etls/requirements.txt
python-dotenv==0.10.3
click==7.0
petl==1.3.0
psycopg2-binary==2.8.3
elastic-apm==5.1.0
cx_Oracle==7.3.0
